python: Rename to framework16_inputmodule
