+++ python/framework16_inputmodule/inputmodule/c1minimal.py
@@ -1,4 +1,4 @@
-from framework16_inputmodules.inputmodule import send_command, CommandVals
+from framework16_inputmodule.inputmodule import send_command, CommandVals
 
 RGB_COLORS = ["white", "black", "red", "green", "blue", "cyan", "yellow", "purple"]
